requirements.txt
aiogram<4.0
python-dotenv==1.0.0
scrapy==2.11.0
sqlalchemy==2.0.25
